change id type from binary into big int (#3)

+++ backend/crates/models/src/topic.rs
@@ -2,19 +2,19 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize)]
 pub struct Topic {
-    pub id: u16,
+    pub id: i64,
     pub topic_name: String,
     pub display_color: String,
     pub created_at: chrono::DateTime<chrono::Utc>,
     pub updated_at: chrono::DateTime<chrono::Utc>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub parent_topic_id: Option<u16>,
+    pub parent_topic_id: Option<i64>,
 }
 
 #[derive(Serialize, Deserialize)]
 pub struct CreateTopicRequest {
     pub topic_name: String,
-    pub parent_topic_id: Option<u16>,
+    pub parent_topic_id: Option<i64>,
     pub display_color: Option<String>,
 }
 
@@ -26,7 +26,7 @@ pub struct UpdateTopicRequest {
 
 #[derive(Serialize, Deserialize)]
 pub struct GetTopicsParams {
-    pub parent_topic_id: Option<u16>,
+    pub parent_topic_id: Option<i64>,
 }
 
 #[cfg(test)]
